requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
g4f>=0.1.0
python-dotenv>=0.19.0